Intelligent Traffic Management at Intersections: Legacy Mode for Vehicles not Equipped with V2V and V2I Communications; L.C.Bento, R.Parafit
Abstract-This paper describes a legacy algorithm for an intelligent
traffic management system applied to automatic regulation of traffic at
intersections. The application of the legacy algorithm enables the
intelligent intersection to accommodate vehicles, in a low percentage,
not equipped or with faulty V2V and V2I communications. The developed
intelligent traffic management techniques, which are based on a
spatiotemporal reservation scheme, aim to minimize accidents, traffic
congestion and consequently the environmental costs of road traffic. The
performance of three intelligent traffic management algorithms applied
to road intersections, roundabouts and crossroads, are analyzed.
Compared with traditional traffic management techniques, simulation
results using the developed intelligent traffic management techniques,
show that the traffic output flux can be increased, traffic flow rate
can be higher and the average time to cross intersections can be
significantly reduced. This reduction is more pronounced when the
traffic flow is heavy. The conducted studies show that the inclusion of
a low percentage of vehicles, not equipped or with faulty V2V and V2I
communications, in intelligent intersections using the legacy algorithm,
have a low impact on the traffic flow.